Update helloWorld.js
I am still very much a newbie to programming and editing open-source and such, but when I compared my exercise solutions to the solutions stored on the repo, I was very confused that this first "Hello, World!" exercise did not seem to have the correct solution in the solutions branch of the repo. In fact, I am still doubting myself, thinking I must be missing something obvious, being such a newbie. But, the code should return "Hello, World!" and not "", right?? Massive apologies if I've missed something obvious, and if so, I am very much trying to learn all I can about web development and would love the feedback. Thanks.